Move and refactor common abstract models and fields
