+++ asso/core/models/commons.py
@@ -1,47 +1,7 @@
 """Base models for other application."""
-
 from django.conf import settings
 from django.db import models
 from django.utils.translation import gettext_lazy as _
-
-
-class SlugModel(models.Model):
-    slug = models.SlugField(
-        unique=True,
-        max_length=100,
-    )
-
-    class Meta:
-        abstract = True
-        ordering = ["slug"]
-
-
-class TitleModel(models.Model):
-    title = models.CharField(
-        max_length=200,
-        default="",
-        verbose_name=_("Title"),
-    )
-
-    class Meta:
-        abstract = True
-        ordering = ["title"]
-
-
-class OrderedModel(models.Model):
-    """A Model with a field for custom ordering."""
-
-    order = models.SmallIntegerField(
-        default=0,
-        null=False,
-        blank=False,
-        verbose_name=_("Order"),
-        help_text=_("Object ordering value"),
-    )
-
-    class Meta:
-        abstract = True
-        ordering = ["order"]
 
 
 class Created(models.Model):
